♻️ llx: digest -> checksum, drop the reflection canon
Rename llx/digest to llx/checksum - checksum is the platform's word for
this everywhere else (code ids, query checksums). Same algorithm, same
AlgoVersion "1", same golden vectors.

Canonicalization is now fully explicit: CanonProto and the protoreflect
walk are gone. AssetValue is hashed field-by-field, and Multiset /
CanonResultMap are exported so cnspec's score/risk checksums fold their
lists and result maps with the exact same encoding instead of reflection.

Features: ScanContentDigestsActive -> ScanContentChecksumsActive; feature
descriptions reworded to checksum. Feature names on the wire unchanged.
