+++ features.yaml
@@ -99,22 +99,22 @@ Features:
   start: v13.x
   status: new
 - id: ScanContentModeShadow
-  desc: "Server-activated scan-content mode (unchanged-scan short-circuit ADR): compute row/kind digests while writing the scandb and stamp them into its metadata. Shadow changes nothing else - every scan still uploads in full."
+  desc: "Server-activated scan-content mode (unchanged-scan short-circuit ADR): compute row/kind checksums while writing the scandb and stamp them into its metadata. Shadow changes nothing else - every scan still uploads in full."
   idx: 18
   start: v13.x
   status: new
 - id: ScanContentModeServerCompare
-  desc: "Server-activated scan-content mode: digests computed and written; the server short-circuits ingest for unchanged uploads. Client behavior is identical to shadow."
+  desc: "Server-activated scan-content mode: checksums computed and written; the server short-circuits ingest for unchanged uploads. Client behavior is identical to shadow."
   idx: 19
   start: v13.x
   status: new
 - id: ScanContentModeClientCompare
-  desc: "Server-activated scan-content mode: digests computed and written, and the client may pull the staged manifest, diff rows, and skip the upload PUT when row-identical, completing with unchanged=true."
+  desc: "Server-activated scan-content mode: checksums computed and written, and the client may pull the staged manifest, diff rows, and skip the upload PUT when row-identical, completing with unchanged=true."
   idx: 20
   start: v13.x
   status: new
 - id: ScanContentModeNoCompare
-  desc: "The kill switch: do NO scan-content digest work and upload in full - behaves exactly like the feature being absent today. Exists as an explicit client-facing signal so that once comparison ships enabled-by-default (no enabling flag), the server can still turn it off per scope; it overrides any other ScanContentMode* feature and any client default."
+  desc: "The kill switch: do NO scan-content checksum work and upload in full - behaves exactly like the feature being absent today. Exists as an explicit client-facing signal so that once comparison ships enabled-by-default (no enabling flag), the server can still turn it off per scope; it overrides any other ScanContentMode* feature and any client default."
   idx: 21
   start: v13.x
   status: new
